backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
django-cors-headers>=4.4,<5.0
psycopg2-binary>=2.9,<3.0
reportlab>=4.1,<5.0
Pillow>=10.0,<11.0
python-decouple>=3.8,<4.0
